Optimized page transitions.
Whole page transition gave the feeling that the website was loading slowly while jumping through pages. This is true due to the animated transition period. Hence, I removed whole page transition and instead made page overlays animate instead of the whole page. The user will see the background picture change instantly and then see the contents in the page "spring down"  giving the feeling of a responsive and alive website.